[DATA] refactor: Organize DATA folder
